requirements.txt
google-search-results
langchain_experimental
pandas==1.4.2
numpy==1.21.5
matplotlib==3.5.1
streamlit==1.24.0
openai==0.27.0
langchain==0.0.141
huggingface-hub==0.11.0
numexpr
python-dotenv
